@@ -16,8 +16,10 @@ internal class AzureServiceBusConfigureOptions : AzureTransportConfigureOptions<
     }
 
     /// <inheritdoc/>
-    public override void PostConfigure(string name, AzureServiceBusTransportOptions options)
+    public override void PostConfigure(string? name, AzureServiceBusTransportOptions options)
     {
+        if (name is null) throw new ArgumentNullException(nameof(name));
+
         base.PostConfigure(name, options);
 
         // ensure we have a FullyQualifiedNamespace when using AzureServiceBusTransportCredentials
