@@ -9,8 +9,8 @@ public class FakeHostEnvironment : IHostEnvironment
 
     public FakeHostEnvironment(string applicationName) => ApplicationName = applicationName;
 
-    public string? EnvironmentName { get; set; }
-    public string? ApplicationName { get; set; }
-    public string? ContentRootPath { get; set; }
-    public IFileProvider? ContentRootFileProvider { get; set; }
+    public string EnvironmentName { get; set; } = default!;
+    public string ApplicationName { get; set; } = default!;
+    public string ContentRootPath { get; set; } = default!;
+    public IFileProvider ContentRootFileProvider { get; set; } = default!;
 }
